+++ mrv2/lib/mrvApp/mrvApp.cpp
@@ -242,7 +242,7 @@ namespace mrv
         _p(new Private)
     {
         TLRENDER_P();
-
+        
         // Establish MRV2_ROOT environment variable
         set_root_path(argc, argv);
 
@@ -1014,7 +1014,7 @@ namespace mrv
         }
 
         for (const auto& path :
-             timeline::getPaths(filePath, pathOptions, _context))
+                 timeline::getPaths(filePath, pathOptions, _context))
         {
             auto item = std::make_shared<FilesModelItem>();
             item->path = path;